Stop passing around neutron client object.
In couple of drivers, in some methods, neutron client object is passing
down to through the series of methods call, just to eventually be
called with some object.

Since we already make the client a singleton, calling
clients.get_neutron_client method will immediately return instance,
which is possibly is already created. There is no need to pass this
down, while it can be easily reached by destination method.

Implements: blueprint switch-to-openstacksdk
Change-Id: I0a08f4740e94f9d7ae5ee003aa8b0a2e0a9b6fac