Adapt dockerfiles to work in both OKD and OCP
A decision was made to force using the same dockerfiles for containers
in both OKD and OCP. As those use two distinct buildsystems and the OKD
one doesn't handle RPMs well, we had the files separated with the OKD
one having a hack to enable repo with RPMs built by the build.

This commit attempts to unify the files in order to use just a single
one. That's done by trying to enable missing repo using sed which will
work for OKD. In order to make it run fine in OCP result of that
operation is simply ignored.

Change-Id: I49a423ffae372b7d8259faf256c68e48324a1b53